fix timestamp cast
